+++ packages/common/src/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './resize';
 export * from './line-path';
+export * from './default-orthogonal-routing';
 export * from './hot-key';
 export * from './creation-mode';
 export * from './direction';
@@ -10,4 +11,3 @@ export * from './memorize';
 export * from './vector';
 export * from './math';
 export * from './drawing';
-export * from './rotate';