+++ packages/draw/src/utils/hit.ts
@@ -8,10 +8,12 @@ import {
     distanceBetweenPointAndSegments,
     distanceBetweenPointAndPoint,
     HIT_DISTANCE_BUFFER,
-    rotate
+    rotate,
+    rotatePoints,
+    isPointInPolygon
 } from '@plait/core';
 import { PlaitDrawElement, PlaitGeometry, PlaitLine } from '../interfaces';
-import { TRANSPARENT, rotatePoints } from '@plait/common';
+import { TRANSPARENT } from '@plait/common';
 import { getTextRectangle } from './geometry';
 import { getLinePoints } from './line/line-basic';
 import { getFillByElement } from './style/stroke';
@@ -63,14 +65,14 @@ export const isRectangleHitDrawElement = (board: PlaitBoard, element: PlaitEleme
             RectangleClient.getCenterPoint(client),
             element.angle
         );
-        return isPolylineHitRectangle(rotatedCornerPoints, client);
+        return isPolylineHitRectangle(rotatedCornerPoints, rangeRectangle);
     }
 
     if (PlaitDrawElement.isLine(element)) {
         const points = getLinePoints(board, element);
         return isPolylineHitRectangle(points, rangeRectangle);
     }
-   return null;
+    return null;
 };
 
 export const isHitDrawElement = (board: PlaitBoard, element: PlaitElement, point: Point) => {
@@ -111,7 +113,13 @@ export const isHitDrawElement = (board: PlaitBoard, element: PlaitElement, point
         }
     }
     if (PlaitDrawElement.isImage(element)) {
-        return isRectangleHitDrawElement(board, element, { anchor: point, focus: point });
+        const client = RectangleClient.getRectangleByPoints(element.points);
+        const rotatedCornerPoints = rotatePoints(
+            RectangleClient.getCornerPoints(client),
+            RectangleClient.getCenterPoint(client),
+            element.angle
+        );
+        return isPointInPolygon(point, rotatedCornerPoints);
     }
 
     if (PlaitDrawElement.isLine(element)) {