+++ packages/draw/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @plait/draw
 
+## 0.51.4
+
+### Patch Changes
+
+-   [#777](https://github.com/worktile/plait/pull/777) [`2a0af6ef`](https://github.com/worktile/plait/commit/2a0af6ef8463b9871f86e311430d6b3efa901811) Thanks [@MissLixf](https://github.com/MissLixf)! - fix(draw): select correct elements with angle
+
+*   [#776](https://github.com/worktile/plait/pull/776) [`3b4be4f0`](https://github.com/worktile/plait/commit/3b4be4f0f4d8e983596a61edb5016bfc6e3919bd) Thanks [@MissLixf](https://github.com/MissLixf)! - fix(draw): fix can not hit image
+
+-   [`c7565c08`](https://github.com/worktile/plait/commit/c7565c08e6d52a91e2df87ac5405bdcd21406cbb) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - rename getPointByVector to getPointByVectorComponent
+    rename getPointByUnitVectorAndDirectionComponent to getPointByVectorDirectionComponent
+
+## 0.51.3
+
+### Patch Changes
+
+-   [#767](https://github.com/worktile/plait/pull/767) [`524cd91e`](https://github.com/worktile/plait/commit/524cd91efc94b8523191fdad95b3579cc4defb03) Thanks [@MissLixf](https://github.com/MissLixf)! - hit select correct with angle
+    rectangle hit select correct with angle
+
+*   [#771](https://github.com/worktile/plait/pull/771) [`1a16e813`](https://github.com/worktile/plait/commit/1a16e81385577ae69f9d0a949809f84ef113e0e1) Thanks [@huanhuanwa](https://github.com/huanhuanwa)! - support addGroup and removeGroup function
+
+-   [#764](https://github.com/worktile/plait/pull/764) [`4df58289`](https://github.com/worktile/plait/commit/4df58289a7818317fe700c803f3dd4f95840e075) Thanks [@huanhuanwa](https://github.com/huanhuanwa)! - support group render and select
+
 ## 0.51.2
 
 ### Patch Changes