+++ packages/text/CHANGELOG.md
@@ -1,5 +1,14 @@
 # text
 
+## 0.51.4
+
+## 0.51.3
+
+### Patch Changes
+
+-   [#767](https://github.com/worktile/plait/pull/767) [`524cd91e`](https://github.com/worktile/plait/commit/524cd91efc94b8523191fdad95b3579cc4defb03) Thanks [@MissLixf](https://github.com/MissLixf)! - hit select correct with angle
+    rectangle hit select correct with angle
+
 ## 0.51.2
 
 ## 0.51.1